fix: Remove payout mode refs (WIP)
